backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
sentence-transformers
google-generativeai
PyGithub
langgraph
pydantic
